refactor: restructure server code by moving functionality to CustomProxyServer and utils
