+++ src/index.ts
@@ -1,3 +1,3 @@
-import { startServer } from './core/server';
+import { startServer } from './server';
 
 startServer();
